automation/requirements.txt
playwright==1.41.0
pandas==2.2.0
openpyxl==3.1.2
requests==2.31.0
python-dotenv==1.0.0
